Move health routes into `health.routes.js`
Our other routes have organically settled on a convention of the routes for a path, for example `/assets`, are all expressed in a matching routes file (`assets.routes.js`).

As we'll be adding many more routes in the future it's a convention that works and will help us manage things. So, we move anything on `/health` into `health.routes.js` and drop the their individual routes files.
