+++ app/plugins/router.plugin.js
@@ -11,12 +11,10 @@
  * @module RouterPlugin
  */
 
-const AirbrakeRoutes = require('../routes/airbrake.routes.js')
 const AssetRoutes = require('../routes/assets.routes.js')
 const BillRunRoutes = require('../routes/bill-runs.routes')
-const DatabaseRoutes = require('../routes/database.routes.js')
 const FilterRoutesService = require('../services/plugins/filter-routes.service.js')
-const InfoRoutes = require('../routes/info.routes.js')
+const HealthRoutes = require('../routes/health.routes.js')
 const RootRoutes = require('../routes/root.routes.js')
 const TestRoutes = require('../routes/test.routes.js')
 
@@ -24,12 +22,10 @@ const AirbrakeConfig = require('../../config/airbrake.config.js')
 
 const routes = [
   ...RootRoutes,
-  ...AirbrakeRoutes,
   ...AssetRoutes,
-  ...BillRunRoutes,
-  ...DatabaseRoutes,
-  ...InfoRoutes,
-  ...TestRoutes
+  ...HealthRoutes,
+  ...TestRoutes,
+  ...BillRunRoutes
 ]
 
 const RouterPlugin = {
