--- app/routes/airbrake.routes.js
@@ -1,17 +1,0 @@
-'use strict'
-
-const AirbrakeController = require('../controllers/health/airbrake.controller.js')
-
-const routes = [
-  {
-    method: 'GET',
-    path: '/health/airbrake',
-    handler: AirbrakeController.index,
-    options: {
-      description: 'Used by the delivery team to confirm error logging is working correctly in an environment. ' +
-        'NOTE. We expect this endpoint to return a 500'
-    }
-  }
-]
-
-module.exports = routes
